[Controller] Implement TAD and improve code (#15)

+++ TTimer.h
@@ -5,13 +5,13 @@
 #include <pic18f4321.h>
 #include "Utils.h"
 
-#define TI_NUMTIMERS 10 // Amount of timers for this TAD, can be modified.
+#define TI_NUMTIMERS 10
 #define NO_TIMERS_AVAILABLE 63
 
 void TiInit(void);
 void Timer0_ISR(void);
 BYTE TiGetTimer(void);
-unsigned int TiGetTics(BYTE timerId);
+WORD TiGetTics(BYTE timerId);
 void TiResetTics(BYTE timerId);
 
 #endif
